refactor: update validation of span
