+++ src/orq_ai_sdk/traced/constants.py
@@ -1,0 +1,2 @@
+# only function is supported for now which then set as SpanType.Generic in the backend - but will support other types especially llm once the POC is completed client sdk + server endpoint
+VALID_SPAN_TYPES = {"llm", "score", "function", "eval", "task", "tool"}
